--- scripts/__init__.py
@@ -1,17 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-# imports
-
-
-__docformat__ = 'reStructuredText'
-
-
-def main():
-    pass
-
-
-if __name__ == '__main__':
-    main()
-
-# EOF
